Move ed25519 tests
